Begun Combat implementation
Added a combat layer to players animator controller. Move to top when testing since the script to control the layer weights (playercombatscript) is not yet functional.

- Added attack1 animation to the controller but have not tested anything with it.

need to figure out how to implement more input actions in the state machine for attack, sheath and draw sword, blocking, change dash to rolls, etc